pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "agentic-task"
version = "0.1.0"
description = "Research Paper Entity Extraction & Citation Analysis Benchmark"
requires-python = ">=3.10"
dependencies = [
    "pandas>=1.5",
    "networkx>=3.0",
    "numpy>=1.24",
    "python-Levenshtein>=0.21",
]

[tool.setuptools.packages.find]
include = ["benchmark*"]
